fix(build): add java home to grudle properties
