+++ user-service/ui/.gitignore
@@ -1,5 +1,6 @@
 *.iml
 .gradle
+/.idea
 /local.properties
 /.idea/caches
 /.idea/libraries
