+++ user-service/ui/app/src/test/java/com/example/user_service/ExampleUnitTest.kt
@@ -1,15 +1,15 @@
-package com.example.user_service
+package com.user_service
 
 import org.junit.Test
 
 import org.junit.Assert.*
 
 /**
- * Example local unit test, which will execute on the development machine (host).
+ * local unit test, which will execute on the development machine (host).
  *
  * See [testing documentation](http://d.android.com/tools/testing).
  */
-class ExampleUnitTest {
+classUnitTest {
     @Test
     fun addition_isCorrect() {
         assertEquals(4, 2 + 2)
